refactor: move find_tempo_challenge to tempo module
Move the Tempo-specific find_tempo_challenge function from
protocol::core::headers to protocol::methods::tempo where it
belongs. Re-export from protocol::core for backward compatibility.

Amp-Thread-ID: https://ampcode.com/threads/T-019d445a-5f14-754e-93cc-84e2bf98afc1
