+++ src/protocol/core/mod.rs
@@ -78,11 +78,14 @@ pub use challenge::{
     PaymentPayload, Receipt,
 };
 pub use headers::{
-    extract_payment_scheme, find_tempo_challenge, format_authorization, format_receipt,
-    format_www_authenticate, format_www_authenticate_many, parse_authorization, parse_receipt,
-    parse_www_authenticate, parse_www_authenticate_all, AUTHORIZATION_HEADER,
-    PAYMENT_RECEIPT_HEADER, PAYMENT_SCHEME, WWW_AUTHENTICATE_HEADER,
+    extract_payment_scheme, format_authorization, format_receipt, format_www_authenticate,
+    format_www_authenticate_many, parse_authorization, parse_receipt, parse_www_authenticate,
+    parse_www_authenticate_all, AUTHORIZATION_HEADER, PAYMENT_RECEIPT_HEADER, PAYMENT_SCHEME,
+    WWW_AUTHENTICATE_HEADER,
 };
+
+/// Re-export from [`crate::protocol::methods::tempo`] for backward compatibility.
+pub use crate::protocol::methods::tempo::find_tempo_challenge;
 pub use types::{
     base64url_decode, base64url_encode, Base64UrlJson, IntentName, MethodName, PayloadType,
     PaymentProtocol, ReceiptStatus,
